django_website/requirements.txt
aiocoap==0.3
aiohttp==3.1.1
appnope==0.1.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
Automat==0.6.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cycler==0.10.0
Cython==0.28.3
dask==0.17.5
decorator==4.3.0
Django==2.0.6
djangorestframework==3.8.2
future==0.16.0
geojson==2.4.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
idna-ssl==1.0.1
imageio==2.3.0
incremental==16.10.1
ipykernel==4.2.2
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
itypes==1.1.0
Jinja2==2.8
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.1
multidict==4.1.0
nbconvert==4.1.0
nbformat==4.0.1
networkx==2.1
notebook==4.0.6
numpy==1.12.1
oauthlib==2.0.6
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.0.2
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.0.7
pyreadline==2.1
pyserial==3.4
python-dateutil==2.4.2
python-debian==0.1.32
pytz==2015.7
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.1.1
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.14.0
scipy==1.1.0
SecretStorage==2.3.1
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sites==0.0.1
six==1.10.0
terminado==0.5
toolz==0.9.0
tornado==4.3
traitlets==4.0.0
Twisted==17.9.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
yarl==1.1.1
zope.interface==4.3.2
django-docs==0.3.0